Change `BitTableBase<>::num_rows_` to `uint32_t`.
For 64-bit targets, changing the member from `size_t` to
`uint32_t` reduces `sizeof(CodeInfo)` from 376 to 336 and
keeps the `Thread::CreateInternalStackTrace()` from reaching
the `-Wframe-larger-than=` limit enforced with `-Werror`.

Test: m

+++ libartbase/base/bit_table.h
@@ -77,7 +77,7 @@ class BitTableBase {
     return table_data_.Subregion(offset, NumColumnBits(column));
   }
 
-  size_t NumRows() const { return num_rows_; }
+  uint32_t NumRows() const { return num_rows_; }
 
   uint32_t NumRowBits() const { return column_offset_[kNumColumns]; }
 
@@ -97,7 +97,7 @@ class BitTableBase {
 
  protected:
   BitMemoryRegion table_data_;
-  size_t num_rows_ = 0;
+  uint32_t num_rows_ = 0;
   uint16_t column_offset_[kNumColumns + 1] = {};
 };
